requirements.txt
asgiref==3.5.2
attrs==21.4.0
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==4.0.6
django-js-asset==2.0.0
django-mptt==0.13.4
factory-boy==3.2.1
Faker==13.15.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==6.2.5
pytest-base-url==1.4.2
pytest-django==4.5.2
pytest-factoryboy==2.5.0
pytest-html==3.1.1
pytest-metadata==2.0.2
pytest-selenium==3.0.0
pytest-variables==1.9.0
python-dateutil==2.8.2
requests==2.28.1
selenium==3.141.0
six==1.16.0
sqlparse==0.4.2
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
